HS-202: code review comments
